Servant.Docs

Description

This module lets you get API docs for free. It lets you generate an API from the type that represents your API using docs:

docs :: HasDocs api => Proxy api -> API

Alternatively, if you wish to add one or more introductions to your documentation, use docsWithIntros:

docsWithIntros :: HasDocs api => [DocIntro] -> Proxy api -> API

You can then call markdown on the API value:

markdown :: API -> String

or define a custom pretty printer:

yourPrettyDocs :: API -> String -- or blaze-html's HTML, or ...

The only thing you'll need to do will be to implement some classes for your captures, get parameters and request or response bodies.

See example/greet.hs for an example.

HasDocs class and key functions

class HasDocs api where

The class that abstracts away the impact of API combinators on documentation generation.

Methods

docsFor :: Proxy api -> (Endpoint, Action) -> DocOptions -> API

docs :: HasDocs api => Proxy api -> API

Generate the docs for a given API that implements HasDocs. This is the default way to create documentation.

docs == docsWithOptions defaultDocOptions

pretty :: Proxy api -> Proxy (Pretty api)

Prettify generated JSON documentation.

docs (pretty (Proxy :: Proxy MyAPI))

markdown :: API -> String

Generate documentation in Markdown format for the given API.

This is equivalent to markdownWith defRenderingOptions.

Customising generated documentation

markdownWith :: RenderingOptions -> API -> String

Generate documentation in Markdown format for the given API using the specified options.

These options allow you to customise aspects such as:

Choose how many content-types for each request body example are shown with requestExamples.

Choose how many content-types for each response body example are shown with responseExamples.

For example, to only show the first content-type of each example:

  markdownWith (defRenderingOptions
                  & requestExamples  .~ FirstContentType
                  & responseExamples .~ FirstContentType )
               myAPI
  

Since: 0.11.1

data RenderingOptions

Customise how an API is converted into documentation.

Since: 0.11.1

Constructors

RenderingOptions

Fields

_requestExamples :: !ShowContentTypes

How many content types to display for request body examples?

_responseExamples :: !ShowContentTypes

How many content types to display for response body examples?

_notesHeading :: !(Maybe String)

Optionally group all notes together under a common heading.

_renderCurlBasePath :: !(Maybe String)

Optionally render example curl requests under a common base path (e.g. `http://localhost:80`).

defRenderingOptions :: RenderingOptions

Default API generation options.

All content types are shown for both requestExamples and responseExamples; notesHeading is set to Nothing (i.e. un-grouped).

Since: 0.11.1

requestExamples :: Lens' RenderingOptions ShowContentTypes

responseExamples :: Lens' RenderingOptions ShowContentTypes

data ShowContentTypes

How many content-types for each example should be shown?

Since: 0.11.1

Constructors

AllContentTypes
For each example, show each content type.

FirstContentType
For each example, show only one content type.

notesHeading :: Lens' RenderingOptions (Maybe String)

Generating docs with extra information

docsWith :: HasDocs api => DocOptions -> [DocIntro] -> ExtraInfo api -> Proxy api -> API

Generate documentation given some extra introductions (in the form of DocInfo) and some extra endpoint documentation (in the form of ExtraInfo.

The extra introductions will be prepended to the top of the documentation, before the specific endpoint documentation. The extra endpoint documentation will be "unioned" with the automatically generated endpoint documentation.

You are expected to build up the ExtraInfo with the Monoid instance and extraInfo.

If you only want to add an introduction, use docsWithIntros.

docsWithIntros :: HasDocs api => [DocIntro] -> Proxy api -> API

Generate the docs for a given API that implements HasDocs with any number of introduction(s)

docsWithOptions :: HasDocs api => Proxy api -> DocOptions -> API

Generate the docs for a given API that implements HasDocs.

newtype ExtraInfo api

Type of extra information that a user may wish to "union" with their documentation.

These are intended to be built using extraInfo. Multiple ExtraInfo may be combined with the monoid instance.

Constructors

ExtraInfo (HashMap Endpoint Action)

extraInfo :: (IsIn endpoint api, HasLink endpoint, HasDocs endpoint) => Proxy endpoint -> Action -> ExtraInfo api

Create an ExtraInfo that is guaranteed to be within the given API layout.

The safety here is to ensure that you only add custom documentation to an endpoint that actually exists within your API.

extra :: ExtraInfo TestApi
extra =
    extraInfo (Proxy :: Proxy ("greet" :> Capture "greetid" Text :> Delete)) $
             defAction & headers <>~ [("X-Num-Unicorns", 1)]
                       & notes   <>~ [ DocNote "Title" ["This is some text"]
                                     , DocNote "Second section" ["And some more"]
                                     ]

data DocOptions

Documentation options.

Constructors

DocOptions

Fields

_maxSamples :: Int

Maximum samples allowed.

defaultDocOptions :: DocOptions

Default documentation options.

maxSamples :: Iso' DocOptions Int

Classes you need to implement for your types

class ToSample a where

The class that lets us display a sample input or output in the supported content-types when generating documentation for endpoints that either:

expect a request body, or

return a non empty response body

Example of an instance:

{-# LANGUAGE DeriveGeneric #-}
{-# LANGUAGE OverloadedStrings #-}

import Data.Aeson
import Data.Text
import GHC.Generics

data Greet = Greet { _msg :: Text }
  deriving (Generic, Show)

instance FromJSON Greet
instance ToJSON Greet

instance ToSample Greet where
  toSamples _ = singleSample g

    where g = Greet "Hello, haskeller!"

You can also instantiate this class using toSamples instead of toSample: it lets you specify different responses along with some context (as ErrorMessage) that explains when you're supposed to get the corresponding response.

Minimal complete definition

Nothing

Methods

toSamples :: Proxy a -> [(Text, a)]

default toSamples :: (Generic a, GToSample (Rep a)) => Proxy a -> [(Text, a)]

toSample :: forall a. ToSample a => Proxy a -> Maybe a

Sample input or output (if there is at least one).

noSamples :: [(Text, a)]

No samples.

singleSample :: a -> [(Text, a)]

Single sample without description.

samples :: [a] -> [(Text, a)]

Samples without documentation.

sampleByteString :: forall ct cts a. (ToSample a, AllMimeRender (ct ': cts) a) => Proxy (ct ': cts) -> Proxy a -> [(MediaType, ByteString)]

Synthesise a sample value of a type, encoded in the specified media types.

sampleByteStrings :: forall ct cts a. (ToSample a, AllMimeRender (ct ': cts) a) => Proxy (ct ': cts) -> Proxy a -> [(Text, MediaType, ByteString)]

Synthesise a list of sample values of a particular type, encoded in the specified media types.

class ToParam t where

The class that helps us automatically get documentation for GET (or other Method) parameters.

Example of an instance:

instance ToParam (QueryParam' mods "capital" Bool) where
  toParam _ =
    DocQueryParam "capital"
                  ["true", "false"]
                  "Get the greeting message in uppercase (true) or not (false). Default is false."

Methods

toParam :: Proxy t -> DocQueryParam

class ToCapture c where

The class that helps us automatically get documentation for URL captures.

Example of an instance:

instance ToCapture (Capture "name" Text) where
  toCapture _ = DocCapture "name" "name of the person to greet"

Methods

toCapture :: Proxy c -> DocCapture

ADTs to represent an API

data Endpoint

An Endpoint type that holds the path and the method.

Gets used as the key in the API hashmap. Modify defEndpoint or any Endpoint value you want using the path and method lenses to tweak.

>>> defEndpoint
"GET" /

>>> defEndpoint & path <>~ ["foo"]
"GET" /foo

>>> defEndpoint & path <>~ ["foo"] & method .~ HTTP.methodPost
"POST" /foo

path :: Lens' Endpoint [String]

method :: Lens' Endpoint Method

defEndpoint :: Endpoint

An Endpoint whose path is `"/"` and whose method is GET

Here's how you can modify it:

>>> defEndpoint
"GET" /

>>> defEndpoint & path <>~ ["foo"]
"GET" /foo

>>> defEndpoint & path <>~ ["foo"] & method .~ HTTP.methodPost
"POST" /foo

data API

Our API documentation type, a product of top-level information and a good old hashmap from Endpoint to Action

apiIntros :: Lens' API [DocIntro]

apiEndpoints :: Lens' API (HashMap Endpoint Action)

emptyAPI :: API

An empty API

data DocAuthentication

A type to represent Authentication information about an endpoint.

Constructors

DocAuthentication

Fields

_authIntro :: String

_authDataRequired :: String

authIntro :: Lens' DocAuthentication String

authDataRequired :: Lens' DocAuthentication String

data DocCapture

A type to represent captures. Holds the name of the capture and a description.

Write a ToCapture instance for your captured types.

Constructors

DocCapture

Fields

_capSymbol :: String

_capDesc :: String

capSymbol :: Lens' DocCapture String

capDesc :: Lens' DocCapture String

data DocQueryParam

A type to represent a GET (or other possible Method) parameter from the Query String. Holds its name, the possible values (leave empty if there isn't a finite number of them), and a description of how it influences the output or behavior.

Write a ToParam instance for your GET parameter types

Constructors

DocQueryParam

Fields

_paramName :: String

_paramValues :: [String]

_paramDesc :: String

_paramKind :: ParamKind

data ParamKind

Type of GET (or other Method) parameter:

Normal corresponds to QueryParam, i.e your usual GET parameter

List corresponds to QueryParams, i.e GET parameters with multiple values

Flag corresponds to QueryFlag, i.e a value-less GET parameter

Constructors

Normal

List

Flag

paramName :: Lens' DocQueryParam String

paramValues :: Lens' DocQueryParam [String]

paramDesc :: Lens' DocQueryParam String

paramKind :: Lens' DocQueryParam ParamKind

data DocNote

A type to represent extra notes that may be attached to an Action.

This is intended to be used when writing your own HasDocs instances to add extra sections to your endpoint's documentation.

Constructors

DocNote

Fields

_noteTitle :: String

_noteBody :: [String]

noteTitle :: Lens' DocNote String

noteBody :: Lens' DocNote [String]

data DocIntro

An introductory paragraph for your documentation. You can pass these to docsWithIntros.

Constructors

DocIntro

Fields

_introTitle :: String

Appears above the intro blob

_introBody :: [String]

Each String is a paragraph.

introTitle :: Lens' DocIntro String

introBody :: Lens' DocIntro [String]

data Response

A type to represent an HTTP response. Has an Int status, a list of possible MediaTypes, and a list of example ByteString response bodies. Tweak defResponse using the respStatus, respTypes and respBody lenses if you want.

If you want to respond with a non-empty response body, you'll most likely want to write a ToSample instance for the type that'll be represented as encoded data in the response.

Can be tweaked with four lenses.

>>> defResponse
Response {_respStatus = 200, _respTypes = [], _respBody = [], _respHeaders = []}

>>> defResponse & respStatus .~ 204 & respBody .~ [("If everything goes well", "application/json", "{ \"status\": \"ok\" }")]
Response {_respStatus = 204, _respTypes = [], _respBody = [("If everything goes well",application/json,"{ \"status\": \"ok\" }")], _respHeaders = []}

Constructors

Response

Fields

_respStatus :: Int

_respTypes :: [MediaType]

_respBody :: [(Text, MediaType, ByteString)]

_respHeaders :: [Header]

respStatus :: Lens' Response Int

respTypes :: Lens' Response [MediaType]

respBody :: Lens' Response [(Text, MediaType, ByteString)]

defResponse :: Response

Default response: status code 200, no response body.

Can be tweaked with four lenses.

>>> defResponse
Response {_respStatus = 200, _respTypes = [], _respBody = [], _respHeaders = []}

>>> defResponse & respStatus .~ 204
Response {_respStatus = 204, _respTypes = [], _respBody = [], _respHeaders = []}

data Action

A datatype that represents everything that can happen at an endpoint, with its lenses:

List of captures (captures)

List of GET (or other Method) parameters (params)

What the request body should look like, if any is requested (rqbody)

What the response should be if everything goes well (response)

You can tweak an Action (like the default defAction) with these lenses to transform an action and add some information to it.

authInfo :: Lens' Action [DocAuthentication]

captures :: Lens' Action [DocCapture]

headers :: Lens' Action [Header]

notes :: Lens' Action [DocNote]

params :: Lens' Action [DocQueryParam]

rqtypes :: Lens' Action [MediaType]

rqbody :: Lens' Action [(Text, MediaType, ByteString)]

response :: Lens' Action Response

defAction :: Action

Default Action. Has no captures, no query params, expects no request body (rqbody) and the typical response is defResponse.

Tweakable with lenses.

>>> defAction
Action {_authInfo = [], _captures = [], _headers = [], _params = [], _fragment = Nothing, _notes = [], _mxParams = [], _rqtypes = [], _rqbody = [], _response = Response {_respStatus = 200, _respTypes = [], _respBody = [], _respHeaders = []}}

>>> defAction & response.respStatus .~ 201
Action {_authInfo = [], _captures = [], _headers = [], _params = [], _fragment = Nothing, _notes = [], _mxParams = [], _rqtypes = [], _rqbody = [], _response = Response {_respStatus = 201, _respTypes = [], _respBody = [], _respHeaders = []}}

single :: Endpoint -> Action -> API

Create an API that's comprised of a single endpoint. API is a Monoid, so combine multiple endpoints with mappend or <>.