Servant.Docs

Description

This module lets you get API docs for free. It lets you generate an API from the type that represents your API using docs:

docs :: HasDocs api => Proxy api -> API

Alternatively, if you wish to add one or more introductions to your documentation, use docsWithIntros:

docsWithIntros :: HasDocs api => [DocIntro] -> Proxy api -> API

You can then call markdown on the API value:

markdown :: API -> String

or define a custom pretty printer:

yourPrettyDocs :: API -> String -- or blaze-html's HTML, or ...

The only thing you'll need to do will be to implement some classes for your captures, get parameters and request or response bodies.

See example/greet.hs for an example.

HasDocs class and key functions

class HasDocs layout where

The class that abstracts away the impact of API combinators on documentation generation.

Methods

docsFor :: Proxy layout -> (Endpoint, Action) -> DocOptions -> API

docs :: HasDocs layout => Proxy layout -> API

Generate the docs for a given API that implements HasDocs. This is the default way to create documentation.

docs == docsWithOptions defaultDocOptions

markdown :: API -> String

Generate documentation in Markdown format for the given API.

Generating docs with extra information

docsWith :: HasDocs layout => DocOptions -> [DocIntro] -> ExtraInfo layout -> Proxy layout -> API

Generate documentation given some extra introductions (in the form of DocInfo) and some extra endpoint documentation (in the form of ExtraInfo.

The extra introductions will be prepended to the top of the documentation, before the specific endpoint documentation. The extra endpoint documentation will be "unioned" with the automatically generated endpoint documentation.

You are expected to build up the ExtraInfo with the Monoid instance and extraInfo.

If you only want to add an introduction, use docsWithIntros.

docsWithIntros :: HasDocs layout => [DocIntro] -> Proxy layout -> API

Generate the docs for a given API that implements HasDocs with with any number of introduction(s)

docsWithOptions :: HasDocs layout => Proxy layout -> DocOptions -> API

Generate the docs for a given API that implements HasDocs.

newtype ExtraInfo layout

Type of extra information that a user may wish to "union" with their documentation.

These are intended to be built using extraInfo. Multiple ExtraInfo may be combined with the monoid instance.

Constructors

ExtraInfo (HashMap Endpoint Action)

extraInfo :: (IsIn endpoint layout, HasLink endpoint, HasDocs endpoint) => Proxy endpoint -> Action -> ExtraInfo layout

Create an ExtraInfo that is garunteed to be within the given API layout.

The safety here is to ensure that you only add custom documentation to an endpoint that actually exists within your API.

extra :: ExtraInfo TestApi
extra =
    extraInfo (Proxy :: Proxy ("greet" :> Capture "greetid" Text :> Delete)) $
             defAction & headers <>~ ["unicorns"]
                       & notes   <>~ [ DocNote "Title" ["This is some text"]
                                     , DocNote "Second secton" ["And some more"]
                                     ]

data DocOptions

Documentation options.

Constructors

DocOptions

Fields

_maxSamples :: Int

Maximum samples allowed.

defaultDocOptions :: DocOptions

Default documentation options.

maxSamples :: Iso' DocOptions Int

Classes you need to implement for your types

class ToSample a where

The class that lets us display a sample input or output in the supported content-types when generating documentation for endpoints that either:

expect a request body, or

return a non empty response body

Example of an instance:

{-# LANGUAGE DeriveGeneric #-}
{-# LANGUAGE OverloadedStrings #-}

import Data.Aeson
import Data.Text
import GHC.Generics

data Greet = Greet { _msg :: Text }
  deriving (Generic, Show)

instance FromJSON Greet
instance ToJSON Greet

instance ToSample Greet where
  toSamples _ = singleSample g

    where g = Greet "Hello, haskeller!"

You can also instantiate this class using toSamples instead of toSample: it lets you specify different responses along with some context (as Text) that explains when you're supposed to get the corresponding response.

Minimal complete definition

Nothing

Methods

toSamples :: Proxy a -> [(Text, a)]

toSample :: forall a. ToSample a => Proxy a -> Maybe a

Sample input or output (if there is at least one).

noSamples :: [(Text, a)]

No samples.

singleSample :: a -> [(Text, a)]

Single sample without description.

samples :: [a] -> [(Text, a)]

Samples without documentation.

sampleByteString :: forall ctypes a. (ToSample a, IsNonEmpty ctypes, AllMimeRender ctypes a) => Proxy ctypes -> Proxy a -> [(MediaType, ByteString)]

Synthesise a sample value of a type, encoded in the specified media types.

sampleByteStrings :: forall ctypes a. (ToSample a, IsNonEmpty ctypes, AllMimeRender ctypes a) => Proxy ctypes -> Proxy a -> [(Text, MediaType, ByteString)]

Synthesise a list of sample values of a particular type, encoded in the specified media types.

class ToParam t where

The class that helps us automatically get documentation for GET parameters.

Example of an instance:

instance ToParam (QueryParam "capital" Bool) where
  toParam _ =
    DocQueryParam "capital"
                  ["true", "false"]
                  "Get the greeting message in uppercase (true) or not (false). Default is false."

Methods

toParam :: Proxy t -> DocQueryParam

class ToCapture c where

The class that helps us automatically get documentation for URL captures.

Example of an instance:

instance ToCapture (Capture "name" Text) where
  toCapture _ = DocCapture "name" "name of the person to greet"

Methods

toCapture :: Proxy c -> DocCapture

ADTs to represent an API

data Method

Supported HTTP request methods

Constructors

DocDELETE
the DELETE method

DocGET
the GET method

DocPOST
the POST method

DocPUT
the PUT method

data Endpoint

An Endpoint type that holds the path and the method.

Gets used as the key in the API hashmap. Modify defEndpoint or any Endpoint value you want using the path and method lenses to tweak.

λ> defEndpoint
GET /
λ> defEndpoint & path <>~ ["foo"]
GET /foo
λ> defEndpoint & path <>~ ["foo"] & method .~ DocPOST
POST /foo

path :: Lens' Endpoint [String]

method :: Lens' Endpoint Method

defEndpoint :: Endpoint

An Endpoint whose path is `"/"` and whose method is DocGET

Here's how you can modify it:

λ> defEndpoint
GET /
λ> defEndpoint & path <>~ ["foo"]
GET /foo
λ> defEndpoint & path <>~ ["foo"] & method .~ DocPOST
POST /foo

data API

Our API documentation type, a product of top-level information and a good old hashmap from Endpoint to Action

apiIntros :: Lens' API [DocIntro]

apiEndpoints :: Lens' API (HashMap Endpoint Action)

emptyAPI :: API

An empty API

data DocCapture

A type to represent captures. Holds the name of the capture and a description.

Write a ToCapture instance for your captured types.

Constructors

DocCapture

Fields

_capSymbol :: String

_capDesc :: String

capSymbol :: Lens' DocCapture String

capDesc :: Lens' DocCapture String

data DocQueryParam

A type to represent a GET parameter from the Query String. Holds its name, the possible values (leave empty if there isn't a finite number of them), and a description of how it influences the output or behavior.

Write a ToParam instance for your GET parameter types

Constructors

DocQueryParam

Fields

_paramName :: String

_paramValues :: [String]

_paramDesc :: String

_paramKind :: ParamKind

data ParamKind

Type of GET parameter:

Normal corresponds to QueryParam, i.e your usual GET parameter

List corresponds to QueryParams, i.e GET parameters with multiple values

Flag corresponds to QueryFlag, i.e a value-less GET parameter

Constructors

Normal

List

Flag

paramName :: Lens' DocQueryParam String

paramValues :: Lens' DocQueryParam [String]

paramDesc :: Lens' DocQueryParam String

paramKind :: Lens' DocQueryParam ParamKind

data DocNote

A type to represent extra notes that may be attached to an Action.

This is intended to be used when writing your own HasDocs instances to add extra sections to your endpoint's documentation.

Constructors

DocNote

Fields

_noteTitle :: String

_noteBody :: [String]

noteTitle :: Lens' DocNote String

noteBody :: Lens' DocNote [String]

data DocIntro

An introductory paragraph for your documentation. You can pass these to docsWithIntros.

Constructors

DocIntro

Fields

_introTitle :: String

Appears above the intro blob

_introBody :: [String]

Each String is a paragraph.

introTitle :: Lens' DocIntro String

introBody :: Lens' DocIntro [String]

data Response

A type to represent an HTTP response. Has an Int status, a list of possible MediaTypes, and a list of example ByteString response bodies. Tweak defResponse using the respStatus, respTypes and respBody lenses if you want.

If you want to respond with a non-empty response body, you'll most likely want to write a ToSample instance for the type that'll be represented as encoded data in the response.

Can be tweaked with three lenses.

λ> defResponse
Response {_respStatus = 200, _respTypes = [], _respBody = []}
λ> defResponse & respStatus .~ 204 & respBody .~ [("If everything goes well", "{ \"status\": \"ok\" }")]
Response {_respStatus = 204, _respTypes = [], _respBody = [("If everything goes well", "{ \"status\": \"ok\" }")]}

Constructors

Response

Fields

_respStatus :: Int

_respTypes :: [MediaType]

_respBody :: [(Text, MediaType, ByteString)]

_respHeaders :: [Header]

respStatus :: Lens' Response Int

respTypes :: Lens' Response [MediaType]

respBody :: Lens' Response [(Text, MediaType, ByteString)]

defResponse :: Response

Default response: status code 200, no response body.

Can be tweaked with two lenses.

λ> defResponse
Response {_respStatus = 200, _respBody = Nothing}
λ> defResponse & respStatus .~ 204 & respBody .~ Just "[]"
Response {_respStatus = 204, _respBody = Just "[]"}

data Action

A datatype that represents everything that can happen at an endpoint, with its lenses:

List of captures (captures)

List of GET parameters (params)

What the request body should look like, if any is requested (rqbody)

What the response should be if everything goes well (response)

You can tweak an Action (like the default defAction) with these lenses to transform an action and add some information to it.

captures :: Lens' Action [DocCapture]

headers :: Lens' Action [Text]

notes :: Lens' Action [DocNote]

params :: Lens' Action [DocQueryParam]

rqtypes :: Lens' Action [MediaType]

rqbody :: Lens' Action [(MediaType, ByteString)]

response :: Lens' Action Response

defAction :: Action

single :: Endpoint -> Action -> API

Create an API that's comprised of a single endpoint. API is a Monoid, so combine multiple endpoints with mappend or <>.